pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-afterimage"
version = "0.3.0"
description = "Episodic memory for Claude Code - persistent memory of code written across sessions with churn tracking"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "DragonShadows1978", email = "noreply@afterimage.dev"}
]
keywords = [
    "claude",
    "claude-code",
    "memory",
    "embeddings",
    "ai",
    "coding-assistant",
    "semantic-search",
    "postgresql",
    "pgvector"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Database",
]
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
embeddings = [
    "sentence-transformers>=2.2.0",
    "torch>=2.0.0",
]
postgresql = [
    "asyncpg>=0.29.0",
    "psycopg[binary]>=3.1.0",
    "pgvector>=0.2.5",
    "numpy>=1.24.0",
    "nest-asyncio>=1.5.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]
all = [
    "ai-afterimage[embeddings,postgresql,dev]",
]

[project.scripts]
afterimage = "afterimage.cli:main"

[project.urls]
Homepage = "https://github.com/DragonShadows1978/AI-AfterImage"
Repository = "https://github.com/DragonShadows1978/AI-AfterImage"
Issues = "https://github.com/DragonShadows1978/AI-AfterImage/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["afterimage*"]

[tool.setuptools.package-data]
afterimage = ["py.typed"]

[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 88
target-version = "py310"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
]
ignore = [
    "E501",  # line too long (handled by black)
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "postgresql: marks tests requiring PostgreSQL",
]

[tool.coverage.run]
source = ["afterimage"]
relative_files = true
omit = ["*/__main__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]
